requirements.txt
asgiref==3.8.1
Django==5.1
django-filter==24.3
djangorestframework==3.15.2
drf-yasg==1.21.7
exceptiongroup==1.2.2
inflection==0.5.1
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
psycopg==3.2.1
psycopg2-binary==2.9.9
pytest==8.3.2
pytest-django==4.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
uritemplate==4.1.1
